refactor(v3/analysis): eliminate circular imports with services layer
Move scoring algorithm functions from services.pr_scoring_service to
analysis.pr_scoring to eliminate upward imports from analysis to services.
This keeps scoring logic co-located with its model types and removes the
last two analysis→services violations.

Changes:
- Move calculate_risk_score, generate_score_report, determine_risk_level
  and helpers from services.pr_scoring_service to analysis.pr_scoring
- services.pr_scoring_service now re-exports for backward compatibility
- Fix inspect_query_service.py to import from analysis.pr_scoring
- Fix pre_push_scope.py to define MAIN_BRANCH_REF locally
- Update analysis/__init__.py to export scoring functions
- Update test imports to use canonical location

Verification:
- All 177 targeted tests pass
- No analysis→services imports remain (rg verified)
- Type checking passes (mypy)
- Backward compatibility maintained via re-exports

Closes #1516
