+++ src/vibe3/analysis/pr_scoring.py
@@ -4,13 +4,17 @@ This module provides pure data models for PR scoring that are shared across
 analysis and services layers. Moving these models to the analysis layer avoids
 bidirectional dependencies between analysis and services.
 
-Note: The scoring logic (generate_score_report, calculate_risk_score) remains
-in services.pr_scoring_service as it depends on configuration and services.
+The scoring logic (generate_score_report, calculate_risk_score) is also located
+here to keep the scoring algorithm co-located with its model types.
 """
 
 from enum import Enum
 
+from loguru import logger
 from pydantic import BaseModel
+
+from vibe3.config.loader import get_config
+from vibe3.exceptions import VibeError
 
 
 class RiskLevel(str, Enum):
@@ -41,3 +45,204 @@ class RiskScore(BaseModel):
     level: RiskLevel
     breakdown: dict[str, int]
     block: bool
+
+
+def _format_trigger_factor(key: str, points: int) -> str:
+    """Format a score contribution for terminal-friendly output."""
+    return f"{key}(+{points})"
+
+
+def _build_reason(score: RiskScore, dimensions: PRDimensions) -> str:
+    """Build a short explanation for the current score."""
+    reasons: list[str] = []
+
+    if dimensions.critical_path_touch:
+        reasons.append("touches critical path")
+    if dimensions.public_api_touch:
+        reasons.append("touches public API")
+    if dimensions.changed_files >= 4:
+        reasons.append("many files changed")
+    if dimensions.changed_lines >= 200:
+        reasons.append("large lines changed")
+    if dimensions.impacted_modules >= 2:
+        reasons.append("wide module impact")
+    if dimensions.codex_verdict == "BLOCK":
+        reasons.append("online review blocks")
+    elif dimensions.codex_verdict == "CRITICAL":
+        reasons.append("online review critical")
+    elif dimensions.codex_verdict == "MAJOR":
+        reasons.append("online review major")
+    elif dimensions.codex_verdict == "MINOR":
+        reasons.append("online review minor")
+    elif dimensions.codex_verdict == "REFUSE":
+        reasons.append("online review refused")
+
+    if not reasons:
+        reasons.append("变更范围较小")
+
+    suffix = "达到阻断阈值" if score.block else "未达到阻断阈值"
+    return f"{'，'.join(reasons)}，{suffix}"
+
+
+def _build_recommendations(score: RiskScore, dimensions: PRDimensions) -> list[str]:
+    """Build actionable recommendations from the scoring dimensions."""
+    recommendations: list[str] = []
+
+    if dimensions.changed_files >= 4 or dimensions.changed_lines >= 200:
+        recommendations.append("拆分 PR 以降低影响面")
+    if dimensions.critical_path_touch:
+        recommendations.append("补充关键路径回归测试")
+    if dimensions.public_api_touch:
+        recommendations.append("明确兼容性影响并补充调用侧验证")
+    if dimensions.impacted_modules >= 2:
+        recommendations.append("补充跨模块联动验证或说明影响范围")
+    if dimensions.codex_verdict in {"MAJOR", "CRITICAL", "BLOCK"}:
+        recommendations.append("先处理审查结论，再继续推进合并")
+    if dimensions.codex_verdict == "MINOR":
+        recommendations.append("可以合并，但建议补 follow-up issue")
+    if dimensions.codex_verdict == "REFUSE":
+        recommendations.append("先澄清审查拒绝原因，再继续推进")
+
+    if not recommendations and score.level in {RiskLevel.LOW, RiskLevel.MEDIUM}:
+        recommendations.append("保持当前粒度，继续按常规测试和审查推进")
+
+    return recommendations
+
+
+def calculate_risk_score(dimensions: PRDimensions) -> RiskScore:
+    """计算 PR 风险分数（从配置读取权重）.
+
+    Args:
+        dimensions: PR 评分维度
+
+    Returns:
+        风险评分结果
+
+    Raises:
+        VibeError: 评分失败
+    """
+    log = logger.bind(domain="pr_scoring", action="calculate_risk_score")
+    log.info("Calculating PR risk score")
+
+    try:
+        config = get_config()
+        w = config.pr_scoring.weights
+        gate = config.pr_scoring.merge_gate
+        t = config.pr_scoring.size_thresholds
+
+        breakdown: dict[str, int] = {}
+
+        # 改动行数
+        cl = dimensions.changed_lines
+        if cl > t.changed_lines.large:
+            breakdown["changed_lines"] = w.changed_lines.xlarge
+        elif cl > t.changed_lines.medium:
+            breakdown["changed_lines"] = w.changed_lines.large
+        elif cl >= t.changed_lines.small:
+            breakdown["changed_lines"] = w.changed_lines.medium
+        else:
+            breakdown["changed_lines"] = w.changed_lines.small
+
+        # 改动文件数
+        cf = dimensions.changed_files
+        if cf > t.changed_files.medium:  # >10
+            breakdown["changed_files"] = w.changed_files.large
+        elif cf >= t.changed_files.small:  # >=4
+            breakdown["changed_files"] = w.changed_files.medium
+        else:
+            breakdown["changed_files"] = w.changed_files.small
+
+        # 影响模块数
+        im = dimensions.impacted_modules
+        if im >= t.impacted_modules.medium:  # >=5
+            breakdown["impacted_modules"] = w.impacted_modules.large
+        elif im >= t.impacted_modules.small:  # >=2
+            breakdown["impacted_modules"] = w.impacted_modules.medium
+        else:
+            breakdown["impacted_modules"] = w.impacted_modules.small
+
+        # 布尔维度
+        if dimensions.critical_path_touch:
+            breakdown["critical_path_touch"] = w.critical_path_touch
+        if dimensions.public_api_touch:
+            breakdown["public_api_touch"] = w.public_api_touch
+        if dimensions.cross_module_symbol_change:
+            breakdown["cross_module_symbol_change"] = w.cross_module_symbol_change
+
+        # Codex 审核结果
+        if dimensions.codex_verdict == "CRITICAL":
+            breakdown["codex_critical"] = w.codex_critical
+        elif dimensions.codex_verdict == "MAJOR":
+            breakdown["codex_major"] = w.codex_major
+
+        score = sum(breakdown.values())
+        level = determine_risk_level(score)
+        block = score >= gate.block_on_score_at_or_above or (
+            dimensions.codex_verdict in gate.block_on_verdict
+        )
+
+        result = RiskScore(score=score, level=level, breakdown=breakdown, block=block)
+        log.bind(score=score, level=level, block=block).success("Risk score calculated")
+        return result
+
+    except VibeError:
+        raise
+    except Exception as e:
+        raise VibeError(f"PR scoring failed: {e}") from e
+
+
+def determine_risk_level(score: int) -> RiskLevel:
+    """根据分数判定风险等级（从配置读取阈值）.
+
+    Args:
+        score: 风险分数
+
+    Returns:
+        风险等级
+    """
+    t = get_config().pr_scoring.thresholds
+    if score >= t.critical:
+        return RiskLevel.CRITICAL
+    elif score >= t.high:
+        return RiskLevel.HIGH
+    elif score >= t.medium:
+        return RiskLevel.MEDIUM
+    return RiskLevel.LOW
+
+
+def generate_score_report(dimensions: PRDimensions) -> dict[str, object]:
+    """生成完整评分报告.
+
+    Args:
+        dimensions: PR 评分维度
+
+    Returns:
+        评分报告 dict
+    """
+    log = logger.bind(domain="pr_scoring", action="generate_score_report")
+    log.info("Generating score report")
+
+    score = calculate_risk_score(dimensions)
+    trigger_factors = [
+        _format_trigger_factor(key, points)
+        for key, points in score.breakdown.items()
+        if points > 0
+    ]
+
+    # Get block threshold from config
+    config = get_config()
+    block_threshold = config.pr_scoring.merge_gate.block_on_score_at_or_above
+
+    report: dict[str, object] = {
+        "score": score.score,
+        "level": score.level,
+        "block": score.block,
+        "block_threshold": block_threshold,
+        "breakdown": score.breakdown,
+        "dimensions": dimensions.model_dump(),
+        "reason": _build_reason(score, dimensions),
+        "trigger_factors": trigger_factors,
+        "recommendations": _build_recommendations(score, dimensions),
+    }
+    log.bind(score=score.score, level=score.level).success("Score report generated")
+    return report
